refactor: extract OverlayUICoordinator from AppModel (Phase 3)
Move overlay/notch UI state machine into a dedicated OverlayUICoordinator,
including notch open/close transitions, notification auto-collapse,
pointer tracking, display configuration, and boot animation. The
coordinator holds a weak AppModel reference for OverlayPanelController
compatibility (which expects the full AppModel). AppModel retains thin
forwarding methods/properties for View and panel controller access.

AppModel: 2117 → 1855 lines (-262)
OverlayUICoordinator: 403 lines (new)
